requirements.txt
gymnasium
numpy
torch
stable-baselines3
wandb
pyyaml
tensorboard
ale-py
opencv-python 
gym[atari]
autorom[accept-rom-license]